% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_detect.R
\name{f_detect}
\alias{f_detect}
\title{Detect factors that should be character or not}
\usage{
f_detect(x)
}
\arguments{
\item{x}{a factor.}
}
\value{
boolean
Returns \code{TRUE} if # unique values = length
An error is raised if # unique values != length
}
\description{
Detect factors that should be character because # unique values = length
}
\examples{
fdetect(factor(c("a", "b", "c","a"))) # FALSE
f_detect(factor(c("a", "b", "c", "d"))) # TRUE
}
