% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_reorder.R
\name{f_reorder}
\alias{f_reorder}
\title{Reorder a factor in descending order}
\usage{
f_reorder(x)
}
\arguments{
\item{x}{a factor.}
}
\value{
Returns a new verision of the input factor in descending order
An error is raised if # unique values != length
}
\description{
a version of reorder() that uses desc() a la (d)plyr
}
\examples{
freorder(factor(c("B", "A", "D"))) # Levels: D B A
}
