% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R, R/write.R
\name{x_read}
\alias{x_read}
\alias{x_write}
\title{Read data frame}
\usage{
x_read(x, x_name, file_path = "new.csv")

x_write(x, x_name, file_path = "new.csv")
}
\arguments{
\item{x_name}{path to store the file of the data frame.}

\item{x}{a data frame.}

\item{file_path}{path to store the companion file.}
}
\value{
\code{x_read()} returns a data frame read from files.
An error is raised if the input is not a dataframe, or files not found.

\code{x_write()} returns nothing.
An error is raised if the input is not a dataframe, or files not found.
}
\description{
\code{x_read()} Functions to read data frames to plain text delimited files
while retaining factor levels

\code{x_write()} Functions to write data frames to plain text delimited files
while retaining factor levels
}
